[CA-13]: Add Home Button functionality to redirect to the home page
Previously, the Home button just a dummy button that only reset the hitCount to 0.

This commit adds the functionality to redirect to the home page when the Home button is clicked.

Thus, the Home button is now a functional button that can be used to redirect to the home page.