[CA-19]: Enable Force Resize for Improved App Screen Adaptability
Previously, the app screen size was not automatically adjusted to fit the screen size of the device.

This commit enables the Force Resize option for improved app screen adaptability.

Thus, the app screen size is automatically adjusted to fit the screen size of the device to ensure that the app seamlessly fits with the screen size, particularly for accurate previews on the Content Management System (CMS).