[CA-30]: Enhance Immersive Mode and Full-Screen WebView
Previously, the app's immersive mode had issues with the visibility of system bars, leaving both the navigation and top bars visible. Additionally, the WebView did not cover the entire display.

This commit addresses these issues by overriding a few superclasses prior to executing the remaining code. It also introduces the use of `useWideViewPort` and `loadWithOverviewMode` settings to ensure that the WebView covers the entire screen.

The changes result in both system bars being hidden, substantially enhancing the immersive mode and ensuring the WebView occupies the entire screen, improving the user experience.