[CA-34]: Fixed scrolling issue in WebView
Previously, the WebView required two-finger scrolling, which can be cumbersome for users.

This commit makes the WebView scrollable with a single finger, by using the `rememberScrollState(0)` instead of `rememberScrollState()` without any initial position. This will allow vertical scrolling with one finger. Also, the `verticalScroll` modifier is removed from the `WebViewComponent`, which was causing the two-finger scrolling behavior.

Thus, it is now possible to scroll the WebView with a single finger, providing a more user-friendly and intuitive experience.