[CA-52]: Implement PasswordInputDialog for security
Previously, the maintenance mode lacked adequate security measures, permitting unrestricted access to anyone. This exposed the app to potential token alterations or access to Android system settings.

This update introduces a `PasswordInputDialog` for the maintenance mode. Although the password is currently hardcoded within the app, the app will eventually receive the password from the server. This addition bolsters security by requiring a password for access to the maintenance mode.

The implementation of the `PasswordInputDialog` secures the maintenance mode, preventing unauthorized access, and ensuring that only authorized personnel can perform maintenance tasks or make changes to the token. This enhancement significantly enhances the app's security posture.