[CA-54]:  Configure WebView User Agent for Improved Compatibility
Previously, the WebView user agent was set to the default value of the Android system. However, some websites may not be compatible with the default user agent.

This change updates the WebView user agent as Desktop.

Thus, change improves compatibility with websites that are not compatible with the default user agent.